Made the display tab dependent on the checkbox
The display tab can be switched on and off independently,
by the checkmark in the toolbox. This is in preparation
of the other display tab.
Also, moved the wait statement to the end of the loop to
minimize the time when the streaming is switched of and the
next read in the same loop.
There are stability problems.