Fixed some threading errors
Previously, on close, the threads kept running. Also, wrote a GUI
update function to prevent threading issues on GUI writes.
Added log messages to beginning and end of each thread.
There are still cases where things get locked either on first connect,
or when the display tab is created. The software now ends with a
crash, which is better than the threads staying alive after the
main window was closed, but should be fixed.