Fixed black screen on Windows, prepared next release
On some graphics cards (Interl UHD 600 and related) on Windows the dancing dots
weren't visible (black screen). This has been fixed.
The version has been updated to 2025.2, for a date based versioning scheme
(who wants to decide when a version update is major or minor, and how many
minors make the next major).