Rippled data display prepared (random data, no display of real data yet)
Implemented a rotating vector containing colours, radii, and positions. Data is represented by
a scalar value per channel. With high frequency, following two things are done:
(1) Each circle's radius is increased by a small fixed value.
(2) for each channel a counter is decreased, proportional to the size of its data value. When the
counter has reached zero, it is reset to one, a new circle with radius zero and position of that
channel is added to the list, while the largest circle is removed.
That way channels with larger values decrease their counters with higher frequency, and circles
will be added faster. Channels with low values will have fewer circles added, or may even remain
totally silent for a while. The effect is like ripples on whater, where active channels cause
many ripples, and inactive channels cause few.