Implemented handing of missing montages and/or channel names
If there are no channel names in the LSL stream, 2d positions are now calculated
based on a rectangular grid. If there are up to five channels that are not part
of an existing montage, the montage is used anyway, and the channels are placed
at the bottom of the layout.
Some minor improvements regarding layout in the curves page (PSD channel positions
adjusted).